Get the PRs for each repo sequentially in the contributors script
There are so many repos in the CanJS org (151 right now) that GitHub started returning rate-limit errors when trying to fetch all the PRs for all the repos at once.

Now, we sequentially fetch the PRs for each repo in an org. This is slower but reliably doesn’t hit GitHub’s rate limits.